+++ fuc/cli/fuccompf.py
@@ -4,8 +4,10 @@ import filecmp
 def create_parser(subparsers):
     parser = subparsers.add_parser(
         get_script_name(__file__),
-        help='[FUC] compare two files',
-        description='This command will compare two files.'
+        help='[FUC] compare contents of two files',
+        description='This command will compare the contents of two files. '
+            "It will return 'True' if they are identical and 'False' "
+            'otherwise.'
     )
     parser.add_argument('file1', help='first file')
     parser.add_argument('file2', help='second file')