+++ fuc/cli/fucexist.py
@@ -7,8 +7,10 @@ def create_parser(subparsers):
         get_script_name(__file__),
         help='[FUC] check whether files/dirs exist',
         description='This command will check whether files/dirs exist. '
-                    'It will look for stdin if there are no arguments (e.g. '
-                    f"$ cat files.list | fuc {get_script_name(__file__)})."
+                    "It will return 'True' if they exist and 'False' "
+                    'otherwise. The command will look for stdin if there '
+                    'are no arguments (e.g. $ cat files.list | fuc '
+                    f'{get_script_name(__file__)}).'
     )
     parser.add_argument('paths', nargs='*',
         help='file/dir paths (default: stdin)')