+++ CHANGELOG.rst
@@ -6,8 +6,10 @@ Changelog
 
 * :issue:`2`: Fix Read the Docs automodule not working properly.
 * :issue:`3`: Add ``sphinx-issues`` extension to Read the Docs.
-* Rename ``fuc.api.BedFrame``, ``fuc.api.FastqFrame``, and ``fuc.api.VcfFrame`` modules to ``fuc.api.pybed``, ``fuc.api.pyfq``, and ``fuc.api.pyvcf``, respectively
-* Add new methods to ``fuc.api.pyvcf`` and ``fuc.api.pyvcf.VcfFrame``.
+* Rename submodules ``fuc.api.BedFrame``, ``fuc.api.FastqFrame``, and ``fuc.api.VcfFrame`` to ``fuc.api.pybed``, ``fuc.api.pyfq``, and ``fuc.api.pyvcf``, respectively
+* Add new methods to ``fuc.api.pyvcf``.
+* Add new methods to ``fuc.api.pyvcf.VcfFrame``.
+* Add new submodule ``fuc.api.snpeff``.
 
 0.1.4 (2021-04-21)
 ------------------